+++ draft-packages/tooltip/docs/Tooltip.stories.tsx
@@ -56,6 +56,17 @@ DefaultKaizenSiteDemo.parameters = {
   },
 }
 
+export const WithNoAnimationDelay = props => (
+  <div
+    style={{ marginTop: "100px", display: "flex", justifyContent: "center" }}
+  >
+    <Tooltip {...props} text="Tooltip label" animationDuration={0}>
+      <Button label="Fast tooltip" />
+    </Tooltip>
+  </div>
+)
+WithNoAnimationDelay.storyName = "With no animation delay"
+
 export const StickerSheet = props => (
   <div
     style={{